+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -39,6 +39,7 @@
 #include "Rule18p7Check.h"
 #include "Rule18p8Check.h"
 #include "Rule19p2Check.h"
+#include "Rule20p2Check.h"
 #include "Rule3p1Check.h"
 #include "Rule3p2Check.h"
 #include "Rule7p1Check.h"
@@ -119,6 +120,8 @@ public:
         "misrac-rule-18p8");
     CheckFactories.registerCheck<Rule19p2Check>(
         "misrac-rule-19p2");
+    CheckFactories.registerCheck<Rule20p2Check>(
+        "misrac-rule-20p2");
     CheckFactories.registerCheck<Rule3p1Check>(
         "misrac-rule-3p1");
     CheckFactories.registerCheck<Rule3p2Check>(