+++ clang-tools-extra/test/clang-tidy/checkers/misrac/rule-20p2.cpp
@@ -1,0 +1,4 @@
+// RUN: %check_clang_tidy %s misrac-rule-20p2 %t
+
+// No FileCheck tests as this would pollute the repository with weirdly named
+// files. Instead, unit tests are provided.